Add session service to keep track of connected sessions

+++ common/src/main/scala/owl/common/OwlService.scala
@@ -5,5 +5,5 @@ trait OwlService {
   def run(): Unit
 
   type Host = String
-  type Port = Short
+  type Port = Int
 }
